refactor: move github utilities and move version logic
Move core/github/operations.py to utils/github_utils.py and update imports.
Inline extract_and_validate_version into core/github/models.py and delete version_utils.py.
Add extract_github_config to core/github/config_service.py and export it from the package. Update tests to use new import paths.
