+++ src/my_unicorn/core/github/__init__.py
@@ -4,13 +4,10 @@ from my_unicorn.core.github.asset_selector import AssetSelector
 from my_unicorn.core.github.client import ReleaseAPIClient
 from my_unicorn.core.github.config_service import (
     GitHubConfig,
+    extract_github_config,
     get_github_config,
 )
 from my_unicorn.core.github.models import Asset, Release
-from my_unicorn.core.github.operations import (
-    extract_github_config,
-    parse_github_url,
-)
 from my_unicorn.core.github.release_fetcher import GitHubClient, ReleaseFetcher
 
 __all__ = [
@@ -23,5 +20,4 @@ __all__ = [
     "ReleaseFetcher",
     "extract_github_config",
     "get_github_config",
-    "parse_github_url",
 ]
