+++ tests/core/github/test_github_ops.py
@@ -2,7 +2,9 @@
 
 import pytest
 
+from my_unicorn.core.github import extract_github_config
 from my_unicorn.exceptions import InstallationError
+from my_unicorn.utils.github_utils import parse_github_url
 
 
 class TestParseGitHubUrl:
@@ -10,7 +12,6 @@ class TestParseGitHubUrl:
 
     def test_parse_github_url_valid(self):
         """Test parse_github_url with valid URL."""
-        from my_unicorn.core.github import parse_github_url
 
         result = parse_github_url("https://github.com/AppFlowy-IO/AppFlowy")
 
@@ -20,7 +21,6 @@ class TestParseGitHubUrl:
 
     def test_parse_github_url_with_trailing_slash(self):
         """Test parse_github_url with trailing slash."""
-        from my_unicorn.core.github import parse_github_url
 
         result = parse_github_url("https://github.com/owner/repo/")
 
@@ -30,7 +30,6 @@ class TestParseGitHubUrl:
 
     def test_parse_github_url_with_extra_path(self):
         """Test parse_github_url with extra path components."""
-        from my_unicorn.core.github import parse_github_url
 
         result = parse_github_url("https://github.com/owner/repo/releases")
 
@@ -40,7 +39,6 @@ class TestParseGitHubUrl:
 
     def test_parse_github_url_invalid_format(self):
         """Test parse_github_url with invalid format."""
-        from my_unicorn.core.github import parse_github_url
 
         with pytest.raises(
             InstallationError, match="Invalid GitHub URL format"
@@ -49,14 +47,12 @@ class TestParseGitHubUrl:
 
     def test_parse_github_url_empty(self):
         """Test parse_github_url with empty URL."""
-        from my_unicorn.core.github import parse_github_url
 
         with pytest.raises(InstallationError):
             parse_github_url("")
 
     def test_parse_github_url_malformed(self):
         """Test parse_github_url with malformed URL."""
-        from my_unicorn.core.github import parse_github_url
 
         with pytest.raises(InstallationError):
             parse_github_url("not-a-valid-url")
@@ -67,7 +63,6 @@ class TestExtractGitHubConfig:
 
     def test_extract_github_config_from_source(self):
         """Test extract_github_config from source."""
-        from my_unicorn.core.github import extract_github_config
 
         effective_config = {
             "source": {
@@ -85,7 +80,6 @@ class TestExtractGitHubConfig:
 
     def test_extract_github_config_missing_source(self):
         """Test extract_github_config with missing source."""
-        from my_unicorn.core.github import extract_github_config
 
         effective_config = {}
 
@@ -97,7 +91,6 @@ class TestExtractGitHubConfig:
 
     def test_extract_github_config_default_prerelease(self):
         """Test extract_github_config with default prerelease value."""
-        from my_unicorn.core.github import extract_github_config
 
         effective_config = {
             "source": {"owner": "test-owner", "repo": "test-repo"}
@@ -111,7 +104,6 @@ class TestExtractGitHubConfig:
 
     def test_extract_github_config_with_prerelease_true(self):
         """Test extract_github_config with prerelease set to True."""
-        from my_unicorn.core.github import extract_github_config
 
         effective_config = {
             "source": {
